refactor: move expectations and aborting modules to "core" folder
