+++ src/core/aborting/errors.ts
