+++ src/core/expectations/classes/string-expectation.ts
