+++ src/index.ts
@@ -1,5 +1,5 @@
 export { createScenario as scenario, mapScenarios } from "@modules/scenarios";
 export { createReusableTest as reusableTest } from "@modules/reusable-tests";
 export { listeners } from "@modules/listeners";
-export { expect } from "@modules/expectations";
+export { expect } from "@core/expectations";
 export { type TestResult } from "@shared/types";
