+++ src/modules/tests/classes/case-test.ts
@@ -1,4 +1,4 @@
-import { TestAborter } from "@modules/test-aborting";
+import { TestAborter } from "@core/aborting";
 import { TestInfo, TestStatus } from "@shared/types";
 
 type Callback = () => void | Promise<void>;
